Safari-609.1.11.1/LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-js-innerhtml-expected.txt

image-set-4x.png has MIME type image/png
PASS fileName(window.testdiv.children[0].currentSrc) is "image-set-4x.png"
PASS successfullyParsed is true

TEST COMPLETE
